Added GitHub Actions workflow
Introduced a new GitHub Actions workflow for building, testing, and publishing. The workflow triggers on pull requests and pushes to the main branch. It includes steps for setting up dotnet, restoring tools and dependencies, building the solution, running tests with coverage reports uploaded to Codecov, linting the solution, packaging the library and uploading artifacts. Additionally, it contains a publish job that runs on push events only and publishes packages to both NuGet.org and GitHub Packages.