Begin to integrate the CE library (#11)
* Added UnMango.Docker.FSharp package to project

The UnMango.Docker.FSharp package has been added as a dependency in the project. This new addition is reflected in the packages.lock.json and fsproj files. The version of this newly included package is 0.0.1-alpha.4, and it also brings along its own dependency on FSharp.Core version 8.0.300-beta.24178.1.

* Added image creation functionality

New features have been added to the Docker.DotNet.FSharp library. The changes include:
- Added new dependencies for Docker.DotNet.Models and UnMango.Docker
- Created a private Convert module to handle image creation parameters
- Extended the Ext type with a Create method that uses the Convert.create function
- Implemented run and runAll functions to execute docker operations sequentially
- Set up a client using DockerClientConfiguration and tested the new image creation feature

* Added project status and image creation example

A brief note about the project's work-in-progress status has been added to the README. Additionally, an example of how to create an image using a minimal computation expression is now included.

* Removed test function from Image module

The test function was removed from the Image module. This function was previously used for creating and running an image, but is no longer needed in the current context.

* Add 'format' target to Makefile for code formatting.

Added a new 'format' target in the Makefile to run code formatting using Fantomas on specified source files.