More container run support (#16)
* Remove debug code

* Updated UnMango.Docker.FSharp package version

The UnMango.Docker.FSharp package version has been updated from 0.0.2 to 0.0.3-alpha.0.1 in the FSharp.fsproj file, ensuring compatibility with the latest updates and features of this package.

* Enhanced container configuration handling

Expanded the functionality of the 'Convert' module in the Container.fs file. Added a new function to handle 'Attach' configurations and significantly extended the existing 'Create' function to support more parameters. Also updated the 'run' function to include these changes.